Move Mercury routes off the edge runtime for static-IP egress
